requirements.txt
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
matplotlib==3.7.1
numpy==1.23.5
PyYAML==6.0
tensorboard==2.13.0
scikit-learn==1.2.2
